+++ R/versions.R
@@ -1,3 +1,24 @@
+##' This function returns the current release and devel Bioconductor
+##' version numbers by parsing the
+##' \code{"http://www.bioconductor.org/js/versions.js"} page.
+##'
+##' @title Bioconductor version numbers
+##' @return A named \code{character} of length two with the
+##'     \code{release} and \code{devel} version numbers (in that
+##'     order).
+##' @author Laurent Gatto
+##' @export
+##' @examples
+##' biocVersions()
+biocVersions <- function() {
+    v <- readLines("http://www.bioconductor.org/js/versions.js")[1:2]
+    v <- gsub("^.+= *\\\"([[:digit:]]+\\.[[:digit:]]+)\\\";", "\\1", v)
+    v <- sort(v)
+    names(v) <- c("release", "devel")
+    v
+}
+
+
 ##' A set of functions to retrieve the version number from packages
 ##' available in Bioconductor or on Github. The version numbers are
 ##' extracted from the Bioconductor package landing page and the