+++ tests/testthat/test_versions.R
@@ -23,3 +23,11 @@ test_that("Package versions", {
     expect_identical(length(ps),
                      length(githubpkgversion(paste0("lgatto/", ps))))    
 })
+
+test_that("Bioconductor version numbers", {
+    v <- biocVersions()
+    ## update me for every new release
+    v0 <- c("3.5", "3.6")
+    names(v0) <- c("release", "devel")
+    expect_identical(v, v0)
+})